Remove duplicate methods
